services/pub-sub-assessment03/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
flask